requirements-dev.txt
-r requirements.txt
flake8==3.7.9
model-bakery==1.0.2
pytest==5.3.1
pytest-cov==2.8.1
pytest-django==3.7.0
